Plugin: moving logic for syntactic tree in object
- Logic to merge trees is now isolated
- Forcing copying of tokens, as they are inferred o/w

NOTE: this will call inferToken, which therefore needs to be implemented. A
temporary hack will be required or we will have to wait for the inference to
be fully implemented.
